test: split into integration and unit tests
